Level 10

+++ todo-app/models/todo.js
@@ -9,52 +9,63 @@ module.exports = (sequelize, DataTypes) => {
      */
     static associate(models) {
       // define association here
+      Todo.belongsTo(models.User,{
+        foreignKey: 'userId'
+      })
     }
 
-    static addTodo({ title, dueDate }) {
-      return this.create({ title: title, dueDate: dueDate, completed: false });
+    static addTodo({ title, dueDate, userId }) {
+      return this.create({ title: title, dueDate: dueDate, completed: false, userId });
     }
     static getTodos() {
       return this.findAll();
     }
-    static deleteTodo(id) {
-      return this.destroy({ where: { id } });
+    static async deleteTodo(id,userId) {
+      return this.destroy({ where: { id,userId } });
     }
-    static async overDue() {
+    static async overDue(userId) {
       return this.findAll({
         where: {
           dueDate: {
             [Op.lt]: new Date().toISOString().split("T")[0],
           },
+          userId,
+          // completed: false,
         },
         order: [["id", "ASC"]],
       });
     }
-    static async dueToday() {
+    static async dueToday(userId) {
       return this.findAll({
         where: {
           dueDate: {
             [Op.eq]: new Date().toISOString().split("T")[0],
           },
+          userId,
+          // completed: false,
         },
         order: [["id", "ASC"]],
       });
     }
-    static async dueLater() {
+    static async dueLater(userId) {
       return this.findAll({
         where: {
           dueDate: {
             [Op.gt]: new Date().toISOString().split("T")[0],
           },
+          userId,
+          // completed: false,
         },
         order: [["id", "ASC"]],
       });
     }
-    static completedItems() {
+    static completedItems(userId) {
       return this.findAll({
         where: {
           completed: true,
+          userId,
         },
+      
         order: [["id", "ASC"]],
       });
     }
@@ -64,9 +75,23 @@ module.exports = (sequelize, DataTypes) => {
   }
   Todo.init(
     {
-      title: DataTypes.STRING,
-      dueDate: DataTypes.DATEONLY,
-      completed: DataTypes.BOOLEAN,
+      title: {
+        type: DataTypes.STRING,
+        allowNull: false, 
+        validate: {
+            notNull: true,
+            len: 5
+        }
+     },
+     dueDate: {
+      type: DataTypes.DATEONLY,
+      allowNull: false, 
+      validate: {
+          notNull: true,
+          len: 5
+      }
+   }, 
+    completed: DataTypes.BOOLEAN,
     },
     {
       sequelize,
